Add logging for missing layout information
The characteristics of the missing bindings in this bug (see
b/324314557#comment14) indicate that the class itself is resolving, but
the fields on the class are not. I'm not sure where that may be
happening, but this change adds logging to the logic that constructs
those fields to help identify some obvious places there may be an issue.

Bug: 324314557
Test: n/a
Change-Id: I9c8c03efcc790e842dae17e7009cb0bcd0c171c1