Fix Build Analyzer warning recommendation link opens twice
The problem comes from the fact that setting URL automatically installs
the listener to open browser. Another listener is added manually
that does both browser and metrics and is tested. So best fix here
is to not trigger setting HyperlinkLabel own listener, set up the right
look of external link manually instead.

Fixes: 328179104
Test: N/A
Change-Id: I1015cdf6a0e0379e1c1be3f141abb552112493c1