Remove _bindingLayoutGroups and lock
Logs from recent canaries have confirmed that bindingLayoutGroups is no
longer accessed in dumb mode, so the logic that cached a copy of the
groups for such a scenario can now be removed. As a result, there's no
locking necessary anymore when fetching the groups, since it relies
entirely on a CachedValue now.

Fixes: 322209412
Test: n/a
Change-Id: I85e5ccd984477483c5b6d9c4dcd365c45fef6b5f