Set task icon order
The order in which the task icons appear has now been
set so that the icons no longer show up in random
order in each Profiler instance.

The ordering is based on a mix of task popularity in
the Sessions-based Profiler, with a slight twist
to push forward the System Trace task as the first
task.

Fixes: 324961100
Test: manual
Change-Id: Id01553ecfc3d15c53d7f9e1be620e3b23c3ca025