Fix tests failing due to "RunClass was already registered"
The issue is basically caused by ActionManagerImpl being instantiated
while the "mask" executors are disposed. When the ActionManagerImpl is
instantiated, it registers some executors which are then attempted to be
re-registered with the call to ExtensionTestUtil.maskExtensions. This
error only happens when there are more than one test case, as a disposal
needs to have occurred before the test case is run. There is a little
bit more information in the linked bug.

The call to ExtensionTestUtil.maskExtensions was actually not necessary
in these tests, as the required executors are already registered as part
of the application.

By removing the call to ExtensionTestUtil.maskExtensions, we avoid
encountering the problem.

Fixes: 329334766
Test: AndroidProfilerLaunchTaskContributorTest,
  AndroidProfilerRunWindowRestorerExecutionListenerTest
Change-Id: I24c92545e9f5982d76ce593f492359e768bde564