Create Quick Definition views in background
With the 2024.1 merge, the construction of
`PsiImplementationViewElement` asserts that it's on a background thread.
Since our quick definition is currently creating it on the event thread,
we need to put that behind a progress indicator.

This is corollary to how the platform handles the exact same case for
their standard `ImplementationViewSession`; see
PsiImplementationViewSession at
https://github.com/JetBrains/intellij-community/blob/master/platform/lang-impl/src/com/intellij/codeInsight/hint/PsiImplementationViewSession.java#L79

Fixes: 329857319
Test: AndroidImplementationViewSessionTest
Change-Id: Ib06733b53af3fbdcbc69b8a38506ed1ad51b8c70