Avoid blocking UI thread to check if device is compatible
Layout Inspector is not supported on emulators with API 29 Play Store
image. We currently check for this all the time, to know if we need to
enabled or disable the action to enable embedded LI.

The check needs to call adb, so it's blocking, calling it from the
action would block the UI thread.

This change modifies the behavior so that the device is considered
to be compatible by default, and an async job is launched to check
if the device is API 29 Play Store image.
This means that the action will be disabled only after the job is
done, so there is going to be a slight delay (less than 1 second).

This is not going to be noticeable by the user. Even if for some
reason they manage to enable Layout Inspector, Layout Inspector
will show a banner explaining why it doesn't work.

Test: existing
Bug: 329714833
Change-Id: I341f77a6b743f46d52c3a3e1588de6a7f485b959