Track activation action enabled state
For deployment targets that are not launched, we need to know if we're
able to launch them before allowing the user to attempt to run a target
on them. For emulators and physical devices, this is not needed because
either they are always launchable or never launchable. However,
Firebase devices may be unavailable momentarily, and need to be
selectable as targets even while they are unavailable.

Thus, we need to incorporate the ActivationAction state into the
LaunchCompatibility for the deployment target.

Bug: 327729426
Test: added
Change-Id: I4b70a94eb3ba00c00bf83916e324f293038e64f0