Remove check for API 29 Play Store image.
Layout Inspector is not supported on emulators with API 29 Play Store
image. We currently check for this all the time, to know if we need to
enabled or disable the action to enable embedded LI.

The check needs to call adb, so it's blocking, calling it from the
action would block the UI thread.

The simplest fix to this issue is to not do the check at all. This is
ok, after enabling Layout Inspector the user will still see a warning
banner explaining why it's not working.

Test: NA
Bug: 329714833
Change-Id: If5c14ab2acfa6b9630af94bb843b59c14d0afbac