Avoid rechecking for consistency of external system data
If data nodes are loaded correctly, and Android models are
also present, we should avoid validating presence of each
individual data node. This change avoids checking if
"IDE_LIBRARY_TABLE" is present, as we should assume that
if android models are present, data has been correctly
deserialized.

Also, this addresses an issue when non-Android Gradle projects
keep on re-syncing on every open.

Bug: 323703308
Test: PlatformIntegrationTest
Change-Id: I76ee9f5bb65c835ae5677c4b1329b3cf2034fa81