Don't use BindingLayoutGroup externally
When fetching LightBindingClasses from the cache, some consumers first
query the cache for BindingLayoutGroups, filter them, and then pass them
back to get the corresponding light classes. This works with the current
storage mechanism, since the light class is store within the
LightBindingClass.

I want to move that storage to address some other bugs, but that
requires that the BindingLayoutGroup passed into getLightBindingClasses
has to be "known". This refactor allows consumers to still filter based
on the group, but it encapsulates that so that the cache can guarantee
the group is one it knows about.

This additionally removes any possibility of a consumer accidentally
passing in a group from a different module, which would result in
non-sensical classes.

Bug: 330744400
Test: Covered by existing
Change-Id: I44dcbd1032d5774fc5d723ed3fcd4847c418959c