Create and store light binding classes in cache
LightBindingClass instances are currently instantiated on demand, and
are stored on their corresponding BindingLayoutGroup. This can cause
problems, since it's not clear whether indexes are available at all
times when the classes may be created.

This change now creates the light classes at the same time as the
corresponding groups. This should not result in excess memory usage as
compared to before; until very recently, BindingScopeEnlarger would
cause all the classes to be instantiated almost immediately upon any
resolution. The effect here is that they'll be created ever-so-slightly
earlier, but the amount of memory usage will be the same.

Bug: 330744400
Test: Covered by existing
Change-Id: Id28cfba8c7d16fd5c51bea2c4c8861e14ac24d16