Add store id to allow to save more zoom levels on surfaces
If we are using different surfaces that are referring to a single
file (for example Layout Editor and Layout Validator) we are
keeping in sync the same scale among the two different surfaces.

This fix adds an additinoal key to the map used to store the scale
to avoid the two scale values are overridden.

Fixes: 248480687
Test: DesignSurfaceSettingsTest
Change-Id: Ia61a350450ff207e06a6fe320aba4bdfc33e71bc