Show a Progress Dialog While Calculating Diff
File-by-file diff takes a long time to calculate. It is already being
executed in a background thread, but it is very CPU intensive so the UI
is not very responsive. Additionally, the background task is not
canceled when we dismiss the analysis popup which makes the UI slow even
after we close the diff.

This change does several things:
1. Displays a modal dialog with a progress bar to prevent the user from
interacting with the UI while the calculation is running.
2. Makes the thread running the task have a low priority.
3. Cancels the task when the progress bar popup is dismissed.

Fixes: 308750490
Test: Manually. There are no existing tests for this component anyway.
Change-Id: I762e75fb1161bdb23041714175773a85d4cc93d5