Fix the UI settings popup location when started from popup toolbar
The relative popup location cannot use the ActionButton, because the
button is in a popup that is closed as part of showing the UI settings
popup.
Convert to AbstractDisplayView coordinates.

Move all balloon related show logic out of the panel and in to a helper
function.

Move tests from the panel to the actions.

Fixes: 331089098
Test: Added unit test
Change-Id: Ied9d0d37884ea20a7245d45722a1b86a54e82b24