Upgrade to Jewel 0.17.3, Compose 1.6.10-dev1584, and Skiko 0.8.1
...by changing module library dependency paths to point to the Jewel
0.17.3, Compose 1.6.10-dev1584, and Skiko 0.8.1 versions rather than
the old, removed 0.13.2, 1.6.0-dev1369, and 0.7.85.4/5 versions.

Some compose dependencies were moved to be shared in KMP, so that
is why some of the *-common dependencies were needed to be explicitly
brought into the compose-desktop module library.

This also brings in the addition of the Jewel markdown artifacts as
part of the jewel-ide module library.

Bug: 272822954
Test: updated
Change-Id: I462422fe269511aad9df62429e038713f5070166