Flatten Past Recordings Tab UX
A new UX design was approved for the past recordings tab where
the recording table takes up the entire tab, and on opening of
a recording, it would take the user to the task tab to choose
what "Viewer" they want to launch the task as. A "Viewer"
is functionally the same as a task in the prior UX.

Nevertheless, the UX change was made, with the exception of the
"Viewer" selection UI, as, after combining the Java/Kotlin Method
Trace and Sample tasks into one task, there is now a 1:1 mapping of
task to importable recording. Thus, a viewer screen would be
pointless at this moment.

Once there is a need (more than one task mapping to a recording),
the "Viewer" selection UI can be added.

Fixes: 332387551
Test: updated
Change-Id: Ic4c8087c381632221656370de0b580188469598f