Allow selecting from TEMP tables in AndroidSQL
b/322575146 points out that the "TEMP" keyword isn't recognized in table
names at the moment. I can't find great documentation on sqlite.org to
directly specify where this can be used, but it's clear from context
that the "TEMP" keyword can be treated as a special database. (See
https://www.sqlite.org/lang_createview.html for an example of the docs
that are helping me infer that.)

The update to the grammar is straightforward, and easy to test with
existing patterns.

All the files under /gen/ were generated by the Grammar-Kit plugin,
version 2022.3.2.

Fixes: 322575146
Test: Included
Change-Id: I94d84c537db45282424bdcc6b15ae451f50d3f79