Remove deprecated exception type in AidlParserTest
JetBrains has made the following change in the 2024.1 update:
https://github.com/JetBrains/android/commit/20176cc6935d6e6bf7d4297959ec61a9c7034ce0

The exception type being caught in this test is deprecated, and they
updated accordingly. That being said, the exception isn't actually used
right now; it's guarded by a flag that's always false. The code in
question is used to more conveniently update golden test sets if
something changes, but isn't actually needed for "real" testing. Given
that this section of code is stable and not changing at this point, I'm
opting to just remove the associated code rather than cherry-pick the\
JetBrains fix.

Bug: 332578858
Test: test code
Change-Id: Ib33fdb684400f06a9c3cc581be8543f8ba2162c6