Queue DumbModeTask in read action
Very occasionally the project may be disposed between our check for it
being disposed and our registering of the DumbModeTask. This leads to
the DumbModeTask being registered as a root disposable (rather than a
child of the project), and it's never disposed.

`Project.isDisposed()` indicates that its value won't change within a
read action, so this should address the rare race condition.

Fixes: 331293193
Test: Covered by existing
Change-Id: I2f8a895ef31e8a5e13ae601fb82719d3a81e8b89