Convert dagger resolution logic to sequences
The primary change here is to use a sequence (instead of intermediate
lists) in DaggerElement.getRelatedDaggerElementsFromIndex,
IndexValue.resolveToDaggerElements, and IndexValue.getResolveCandidates.
All of the other changes are just updating returned lists to be
sequences instead.

This has two benefits; the first (suggested by an intention in
DaggerElement.kt) is that it avoids creating a ton of intermediate
lists. Given the amount of filtering/transforming going on, the wrapping
sequences should be a lot lighter-weight than all those lists. Second
(and importantly for the freezing bug), this makes it easier to put in
process cancellation checks that will be run as each individual item is
resolved.

Bug: 327493766
Test: Included
Change-Id: Ib2c4d5a823b23e6539e6251aa9a20bd9fd217a34