Cache related dagger items
In case the process is cancelled while gutter icons are being generated,
this cache will help save the results that have already been created
once.

Putting this behind a flag so that if for some reason the caching
introduces problems we have a way to quickly disable it (especially for
ASwB). I don't expect this; but one way it could happen is if there's
any possibility for a single PsiElement to be interpreted as multiple
types of DaggerElements. That should theoretically not happen, but I
believe this is the first piece of logic that would actually depend upon
that precondition (and if it fails, we may get a consistency error for
the cached value). So the flag is largely just a safety mechanism at
this point.

Fixes: 327493766
Test: n/a
Change-Id: I9a1ba6a9bebdbd9416f06ebb39728ecfb7c71763