[Dsl] Shard gradle-dsl tests
The test suite duration is a (high, particularly on Windows) setup
time plus a small time (around 100ms) multiplied by the number of
tests.  The number of tests scales as the number of Dsl languages
supported; we are currently at 4, which is enough to push the
occasional run of the target on Windows postsubmit above the "long"
threshold of 15 minutes.

By sharding the gradle-dsl tests to the same number of shards as the
number of Dsl languages, we should make the time taken by any
individual shard approximately independent of the number of Dsl
languages, at the of paying the setup time N times rather than 1.

Bug: 332569757
Test: monitor timeouts of this test target
Change-Id: Ie458358862f62ca297e3c2d7b727b6fe2045c7e2