[Dsl] Disable leak checker on Gradle Dsl tests
There is a race condition between service initialization and
registering the service in the disposer tree with the correct parent.
As a result, if we close the project between service initialization
and the registration, we leak the service.

Probably for most services, this is harmless.  But
XDebuggerManagerImpl registers itself as the disposable for a project
message bus connection, which consequently itself never gets disposed,
and this triggers the LastInIdeaTestSuite leak checker test.

For now, disable the leak checker on Gradle Dsl tests, but really the
race window in the platform is problematic and should be closed as
much as possible: or at least the message bus connection should be
explicitly parented.

Bug: 333499915
Test: existing
Change-Id: Ib57d310a3390c94b23d0fb194688b8f093498565