Refactor Profiler icons to use StudioIconsCompose
With the addition of StudioIconsCompose providing the compose-based
icons as ResourcePainterProviders, there is now no need to define
our own icon structure (e.g. TaskBasedUxIcon) and now we can fetch
the icon directly from StudioIconsCompose.

This change moves all usages of TaskBasedUxIcon over to use
StudioIconsCompose. It also removes the declarations of unused
icons.

This fixes the issue that the classic UI icon assets were being
retrieved rather than the assets for the new UI.

Bug: 277797248
Test: manual
Change-Id: If2a8613853622cf49a862f2fe6345b79292af65e