Gracefully handle TaskArgs/task launching illegal state exception
Currently, if there is an issue creating the task arguments within
the lanuchProfilerTask method, an IllegalStateException is thrown.
This exception is thrown and crashes the profiler, as in, the user
can no longer perform any tasks after it is thrown once.

This change adds exception handling to gracefully handle such error,
and does so by catching it, then displaying a error dialog to the
user.

This change does not address why the tagged bug is throwing the
original exception, but it does prevent the symptom of the profiler
crashing.

Bug: 333951152
Test: manual
Change-Id: Ie52546f189689c0317ca2143e957365734beadaf