Fix CaptureDetailsTest flaky test
Exception occurs while expanding rows in Jtree node. Wrap with
"SwingUtilities.invokeAndWait" method while trying to add nodes
in tree in CaptureDetailsTest class.

SwingUtilities.invokeAndWait is a utility method provided by Swing
to execute a piece of code asynchronously on the Event Dispatch
Thread (EDT). The expansion of the nodes was happening in main
thread for tests and now happens on EDT.

Bug: 317126560
Test: Modified
Change-Id: If7dc1ab3ad475eebf1eb7a1d787a35effa00b669