Fix race between foreground process detection and app inspection
Sometimes a race condition happens between foreground process
detection and app inspection process discovery, where the
foreground process is detected before the process is available
in app inspection.

This change fixes this bug by adding an observer on app inspection
processes. When a process shows up, and is the current foreground
process, we connect to it.

Test: ForegroundProcessDetectionInitializerTest
Bug: 335413000
Change-Id: I02e3cc571ee10fbe7c89b8afb61288fb8df232bb