Register disposable only when it is created
This change attempts to address b/333932711. It appears that
the problem is when addLibrary is called multiple times
with the same "repo" argument. In this case, teh call to
"getOrPut" will use the "get" path, but we will register
the disposable that cleans it up an additional time.

Instead, only register the disposable the first time the
"repo" is added as a key to the "repoMap".

Fixes: 333932711
Test: ResourceClassRegistryTest.kt
Change-Id: Ic25ac8f60932efda93d0bbc72ecc7ccc0b87f840