Fix Live Edit prebuild checks to handle unsupported test source
The TestArtifactSearchScopes is only responsible for determining the
test scope for android-specific modules. Meaning that in case of KMP
project, isn't able to properly determine if the modified file is
a test source or not for example for commonMain.

For this reason, it's necessary to also validate if is a test source
using the platform-specific API.

Bug: 319247142
Test: Existing
Change-Id: I421073151dcb84991649a5174236fe22353213b8