Create action to generate Compose Previews for composables
The action builds a prompt that explains an LLM how to create a Compose
Preview corresponding to a simple composable. Then, it asks the model
to generate a Compose Preview corresponding to the composable that's
being right-clicked.

When the model query is received, we show it in a diff view, so users
can accept it entirely or partially.

Bug: 335418850
Test: TBD
Change-Id: I035b9fe78427a4ce78602e950ff8755728da6577