Fix "show issue explanation" rendering for code blocks in the IDE
If you bring up the "show more" panel for an issue in the IDE, and that
explanation contains a fenced code block, then the lines in the code
block are all concanated into a single line!

This happens because the HTML renderer used in inspections popups
has some strange behaviors; in particular, it renders \n's as newlines
(which is not what is supposed to happen in HTML), so the code was
working around this by replacing them all, since there are generally
<br>'s where we need newlines. But there are no <br>'s in <pre> tags,
so these lines ended up getting joined.

Test: Unit test included
Bug: Not filed
Change-Id: I3ed7ef758501ee15af1b430ddeeb82dd9c395337