Run open file action in application scope
This action was previously running in a project scope associated with
the project being closed (as another project opened). That project's
jobs would be cancelled, causing the opening of the new project to be
cancelled.

It looks like this was probably introduced with
https://github.com/JetBrains/intellij-community/commit/edb9bde0. Prior
to that change, it seems that all actions defaulted to running in an
application scope if they didn't specify anything (see
`AnAction.getActionPerformScope()` and `ActionManagerImpl` in that
change). Following the change, it looks like
`templatePresentation.isApplicationScope` defaults to false, so this may
affect other actions as well.

Fixes: 335723195
Test: Manually verified
Change-Id: I86cb63b3adb18a97672041367a87c2508bacd112