Remove download source button by unregistering extension
For Android libraries, Download Source button wouldn't make a
difference if we/AS failed to download the sources(either because
source doesn't exist or the library is multiple variant library).
However, users tend to click that button and expect it will download
a source for them, which ends up a confusing error for them. For java
libraries, we/AS have configured on code level to instruct intellij
to download the sources automatically, so users in theory won't need
to click that button.

Bug: 327665611
Test: DownloadSourceTest
Change-Id: Ic0fa99ec54e6612b936a2d87ea2730ee52fb9170