Add warning + errors for task home tab selections
1. Refactor the selection verification to be outside the
   data model and into static methods. These methods are
   shared with a method that finds the selection errors.

2. Add method to detect if the user is using a profileable
   preferred task with a non-profileable process.

Using (1) and (2), the warning or error message UI was added
to the bottom bar of the task home tab.

Because this change introduces more content into the task
action bar, this change also includes some UI tweaks to make
the UI responsive to smaller width displays. For example,
if text is overflown and showing ellipses, a tooltip will
appear with full content of the text. Another UI tweak for
responsiveness is the collapse of the added warnings/errors
into just an icon if the width is below a threshold width.

Per the UX spec, when the detection of (2) is made, the text
of the "Start profiler task" button is changed to be
"Start anyways".

Fixes: 315033624
Fixes: 326627234
Fixes: 309506699
Bug: 277797248
Test: updated + added
Change-Id: I88e0f0ec8616f5e8c2e25e7650b3ea58a2d5606e