Relax creating the GradleWrapper in tests a little
If we cannot detect the embedded repository, do not update the
distributionUrl property from what was created.

This allows us to remove Gradle artifacts from the test dependencies
of the catch-all __other split test target in Gradle project system
tests, the rationale being that any test which actually needs to run
Gradle (and hence be sensitive to the contents of the Gradle wrapper
file) should be in a different split target.

Bug: 333945762
Test: existing
Change-Id: I8bffe3dbbcbe79874d18c1f4885f1386dd0123ed