Add display config for autmotive devices
Added hw.display6 configurations for the devices which supports cluster
display and hw.display7 configurations for the automotive distant
display devices. Those values would be referred from studio when
updating display configration on snapshot loading.
This is a temporal resolution and needs to be removed when the runtime
display configuration is ready.

Bug: 337271836
Bug: 335781727
Test: build local androd studio and confirm the config.ini for
automotive devices

Change-Id: Ie85f77b7077d242b9ea92d90a6ed19b886a0bbac