Split TransportDeivceManagerListener in two interfaces
This change splits TransportDeviceManagerListener in to interfaces:
1. TransportDeivceManagerListner, which contains methods to listen
to events from the Transport.
2. TransportConfigContributor, which contains methods used by clients
of the transport to customize its configuration before initialization.

TransportConfigContributor is made into an extension point.

By having an extension point, different clients of the Transport
can provide their configuration to the daemon, even if they have
not yet been registered to the topic.

Test: existing
Bug: 337816875
Change-Id: Idade687f67b64e86d40bdd78934f795a683392b2