Move javac deprecation warnings code to a separate parser
Make logic for parsing java language level deprecation warnings and
errors to be a separate parser rather than part of AGP output parser.

With new deprecation of 7 in java 21, and new message in AGP for it,
we would like to change this warning presentation. Parsing will need to
become more complex, it needs to be implemented as a separate parser.

For now copy logic from AndroidGradlePluginOutputParser, refactoring and
simplifications will follow with new quickfixes implementation.

Also making the related test to have stronger assertions about what
issues are generated for build output window.

Bug: 327626363
Test: JavaLanguageLevelDeprecationOutputParserTest
Change-Id: I3e707ff85cc6633f66be06c66a90faaeb329bda3