Stop generating safe arg classes for scope
Similar to ag/26518892, this change will modify the search scope for
safe args light classes to look for a predefined data key, rather than
requiring all the light classes to be enumerated in order to create the
scope. The search scope infrastructure calls into this logic in dumb
mode, which has the potential to cause stale values to get cached.

The test being removed is no longer relevant, since the search scope
doesn't care anymore about dumb mode. The remainder of the logic is
tested in depth by numerous safe args tests; any test that requires
resolving the safe args classes (of which there are many) fails if the
search scope isn't set up correctly to identify these classes.

Bug: 338060236
Test: Covered by existing
Change-Id: I25363b993c672f508fc4ee379e36044b8a558a99