LE: Bail on non-compose modules
The Compose Compiler would complain when it does not sell the
Compose runtime in the module's dependencies. It will then throw
an exception which makes it look like an internal error in the metrics.

This CL basically prevents this case from happening.

We can think about how we can support this by invoking the normal
Kotlin compiler.

Fixes: 313839326
Test: testNonComposeModule
Change-Id: Ib6892d3594d4b7a0f7ab52ef6addc9ddb5f40425