Add flag to indicate when the thread should terminate
The TransportThread should be terminated when the adb bridge changes.
Currently this is done by cancelling the future returned by the
executor that runs the thread.

Future cancellation occasionally fails, because before receiving the
interrupt the thread can invoke myDevice.executeShellCommand. After
which the thread blocks and does not check for cancellation anymore.

By passing down a flag that indicates when the thread should stop
we guarantee that we can always recover the thread, even if
myDevice.executeShellCommand is called before the interrupt is received.

Test: NA
Bug: 331183470
Change-Id: I7088caec8cb6c9cbbb7d09d37690a6c1f47739d0