Refactor BasePreviewAnnotationInspection
This change mainly extracts some of the abstract methods of this
class into an interface. As a consequence, this part of the class
will be configurable when creating instances, without neeeding to
define a subclass. And this will allow for many other existing
inspections to be reused by Glance Preview in upcoming CLs.

Bug: 338236819
Test: existing
Change-Id: If16907b8ade32fdbbc4e29811f70189ed4f1a200