Move tflite sample code creation to background
The tflite editor already has a mechanism where it doesn't display
Sample Code right away. Rather, it waits for a notification that
indexing is complete before it does that, and then adds in Sample Code
after that.

This change moves the generation of sample code to the background, to
avoid potential disk access on the EDT. This means it may take slightly
longer for the code to display, but that paradigm already exists in the
UI so it shouldn't be a jarring experience.

Fixes: 338362764
Test: TfliteModelFileEditorTest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:b53ed4feee350384f9242e1945916a3fcab0f714)
Merged-In: Ib492a65179ad157238e6d0af8bbff0ea41670ece
Change-Id: Ib492a65179ad157238e6d0af8bbff0ea41670ece