Fix VisualizationForm freeze
When initializing VisualizationForm, since it mostly runs in the UI
thread, addModelWithoutRender results should not be waited on in the UI
thread or can cause freezes.

This change moves from the join call to simply use the common
CompletableFuture stages to complete.
join() would freeze in the cases where something is scheduled to run in
the UI thread but the join is actually happening in the UI thread.

With this change, there is no wait and the control is correctly handed
over from thread to thread.

Test: VisualizationFormTest
Fixes: 338510256
Change-Id: I0668f5bcc695eccdf7bf42396005a08dba8abd4a
(cherry picked from commit 6116c4d406d517ef507e01b75a36e00a2e0c18b3)