Create style reference for whole style
There is special logic for creating multiple references within a style
name when the name has one or more parents. However, this logic does not
create a reference for the entire name itself.

Usually what happens is this logic returns an empty array, and so
default logic higher up runs and creates a reference for the entire
name. In this special case, that final reference isn't being created
higher up because the returned results aren't empty.

Adding this reference allows searching of find usages on the `A` and `B`
parts of the name referenced in b/235848109 separately.

Fixes: 235848109
Test: AndroidResourcesFindUsagesTest.testStyleInheritance3, AndroidValueResourcesTest
Change-Id: Id8f1ead144d455cb42abc94dde9a8835574e7925