Migrate LintIdeFixPerformer to ModCommand API
The interesting part of the change is introducing a FileProvider
interface, with read-only and read-write implementations on the IDE
side. This is because constructing the LintIdeFixPerformerFix
requires access to the files in order to build smart Psi ranges, but
the actual writes should all happen during intention preview or actual
invocation, both within the context of a ModPsiUpdater mediating the
file accesses (calls to getWritable, including to the parent
PsiDirectory when creating a new file, etc).

There's no support for binary files yet, but we also don't seem to
have quick fixes that use this feature either, so I've left a TODO
and will follow up with upstream.

The InspectionQithQuickFixesTest integration test verifies the Lint
external annotator -> LintFix -> LocalQuickFix pipeline actually
works. It was making this test pass that prompted the change to how
we register the intention in LintExternalAnnotator to make sure we
provide the correct key. As a result, now these quick fixes have the
extra side menu to include navigating to inspection settings etc. just
like IntelliJ's.

Bug: 319287252
Test: existing
Change-Id: Ie9249f65c099b7796df7b06f84681c5e825d95f1