Prevent faulty tasks recordings from showing in list
By adding the intended task the user initiates to the
session metadata, we can now detect faulty task recordings.
These faulty task recordings typically fail to capture any
underlying artifact, and because live view recordings also
had no artifact, they were detected as live view recordings.
But now, if we know the user inteded to do a, for example,
system trace task, but no artifact was produced, we know
it was not a live view, but rather a faulty recording.

These faulty recordings are now filtered out of the past
recordings list.

Bug: 277798530
Test: added + updated
Change-Id: Iac200b2f003fe2428c0eed295b950a2f4c7cac5d