Fix device filter and url
We stopped including the manufacturer as a prefix to
the device name in I2cedfe5c2, but it's required for
the device filter. Adding it back there, as well as in
the play console URL.
Also fixed a bug where the name of the device filter
param in the play console URL was incorrect.

Fixes: 339788186
Test: updated. This still isn't great, since it relies on us
correctly mimicing upstream behavior.

Change-Id: Ie04c001e74faaa4eb056519d632a117bb3628d17