Get Rid of LaunchableAndroidDevice
It is being used in only 2 places. One of them is a single test class.
The other is a bug. It's being used to determine if we want to
 `Record stat if we launched a device` but this doesn't work anymore.
The proper way to test for that is to use `!AndroidDevice.isRunning`
which is the way it's done in `AndroidRunConfigurationBase`.

Since the only remaining usage in a test module, I move the class to
that module and renamed it to reflect that it's not a production class.

Test: NA
Bug: NA
Change-Id: I0d415fa84b03cc93d3f069fc31b1729dc8709248