Add build action for non gradle projects doing startup tasks
When a user attempted a startup task with a non-Gradle project,
no build action was kicked off, and thus the startup task would
fail. This change introduces a third action (ProfileAction) that
is used to build and launch the app on invocation of a startup
task when using a non-Gradle project.

This change also moves the profiler actions one directory up from
the 'actions' directory into the 'actions' directory.

Fixes: 341790104
Test: manual
Change-Id: I60d22b3e3b71f3dd1cebf750c955e0ba032eb8d7